Decomposed Media plus DDL changes
I decomposed Media into Media_Content and Media. The endpoint usage should be the same except for a small tweak. When you go to update a form of media, you cannot directly change the title of it since it is used as a primary key. Instead, I made a new endpoint "media/update_title" that takes in the new title and media ID. That will update the title of a form of media if you want to. I also changed usernames to being unique, made phone numbers strings, removed rent prices since everything is free now.